requirements.txt
Pillow
PyPDF2==1.26.0
pdf2image==1.16.0